+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 1.0.0-rc.2 - 23 August 2016
+
+###### Backwards compatible changes
+- Add support for `field` option to `findAll`.
+
 ##### 1.0.0-rc.1 - 16 August 2016
 
 - Initial Release